chore(source): check schema on table source writing chunks
