+++ src/common/src/util/mod.rs
@@ -34,6 +34,7 @@ pub mod try_match;
 pub mod epoch;
 mod future_utils;
 pub mod scan_range;
+pub mod schema_check;
 pub mod sync_point;
 
 pub mod value_encoding;
