+++ src/source/src/table_v2.rs
@@ -69,6 +69,13 @@ impl TableSourceV2 {
                 .clone()
         };
 
+        #[cfg(debug_assertions)]
+        risingwave_common::util::schema_check::schema_check(
+            self.column_descs.iter().map(|c| &c.data_type),
+            chunk.columns(),
+        )
+        .expect("table source write chunk schema check failed");
+
         let (notifier_tx, notifier_rx) = oneshot::channel();
         tx.send((chunk, notifier_tx))
             .expect("write chunk to table reader failed");
